Migrated some mercAllocator.cpp implementation to mercAllocator.hpp
